chore: check code specification
