--- flybirds/core/plugin/plugins/default/__init__.py
@@ -1,2 +1,0 @@
-from .web.screen import Screen
-from .web.screen_record import ScreenRecordInfo
