+++ flybirds/utils/file_helper.py
@@ -28,7 +28,7 @@ def get_json_from_file_path(path):
         f = open(path, "r", errors='ignore')
         json_str = f.read().strip().replace('\r\n', '\\r\\n')
         result = json.loads(json_str, strict=False)
-    except JSONDecodeError as e:
+    except JSONDecodeError:
         log.warn('get_json_from_file_path has error.')
     finally:
         if f:
